requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-embed-video==1.3.3
django-js-asset==1.2.2
gunicorn==20.0.4
idna==2.10
Pillow==7.2.0
psycopg2==2.8.5
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
